sFileCore version 1.0.0

+++ lib/readFile.h
@@ -2,6 +2,6 @@
 #define READFILE_H
 
 char *readFile(char fileName[], int stringSize);
-int readValue(char fileName[], char key[], int stringSize);
+float readValue(char fileName[], char key[], int stringSize);
 
 #endif
